Renamed coordination component
