+++ components/org.wso2.carbon.stream.processor.core/src/main/java/org/wso2/carbon/stream/processor/core/impl/HaApiServiceImpl.java
@@ -21,10 +21,10 @@ package org.wso2.carbon.stream.processor.core.impl;
 import org.apache.log4j.Logger;
 import org.wso2.carbon.stream.processor.core.api.*;
 import org.wso2.carbon.stream.processor.core.api.NotFoundException;
-import org.wso2.carbon.stream.processor.core.coordination.HACoordinationSinkHandler;
-import org.wso2.carbon.stream.processor.core.coordination.HACoordinationSinkHandlerManager;
+import org.wso2.carbon.stream.processor.core.ha.HACoordinationSinkHandler;
+import org.wso2.carbon.stream.processor.core.ha.HACoordinationSinkHandlerManager;
 import org.wso2.carbon.stream.processor.core.internal.beans.ActiveNodeLastPublishedEventTimeStamp;
-import org.wso2.carbon.stream.processor.core.coordination.util.CompressionUtil;
+import org.wso2.carbon.stream.processor.core.ha.util.CompressionUtil;
 import org.wso2.carbon.stream.processor.core.internal.SiddhiAppData;
 import org.wso2.carbon.stream.processor.core.internal.StreamProcessorDataHolder;
 import org.wso2.carbon.stream.processor.core.model.HAStateSyncObject;
