+++ components/org.wso2.carbon.stream.processor.core/src/test/java/org/wso2/carbon/stream/processor/core/HACoordinationSinkHandlerTest.java
@@ -24,8 +24,8 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.testng.PowerMockTestCase;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import org.wso2.carbon.stream.processor.core.coordination.HACoordinationSinkHandler;
-import org.wso2.carbon.stream.processor.core.coordination.HACoordinationSinkHandlerManager;
+import org.wso2.carbon.stream.processor.core.ha.HACoordinationSinkHandler;
+import org.wso2.carbon.stream.processor.core.ha.HACoordinationSinkHandlerManager;
 import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.stream.output.sink.PassThroughSinkMapper;
 import org.wso2.siddhi.core.stream.output.sink.SinkHandlerCallback;
