+++ components/org.wso2.carbon.stream.processor.core/src/test/java/org/wso2/carbon/stream/processor/core/HACoordinationSourceHandlerTest.java
@@ -23,9 +23,9 @@ import org.powermock.core.classloader.annotations.PrepareForTest;
 import org.powermock.modules.testng.PowerMockTestCase;
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import org.wso2.carbon.stream.processor.core.coordination.HACoordinationSourceHandler;
-import org.wso2.carbon.stream.processor.core.coordination.HACoordinationSourceHandlerManager;
-import org.wso2.carbon.stream.processor.core.coordination.util.CoordinationConstants;
+import org.wso2.carbon.stream.processor.core.ha.HACoordinationSourceHandler;
+import org.wso2.carbon.stream.processor.core.ha.HACoordinationSourceHandlerManager;
+import org.wso2.carbon.stream.processor.core.ha.util.CoordinationConstants;
 import org.wso2.siddhi.core.event.Event;
 import org.wso2.siddhi.core.stream.input.source.InputEventHandler;
 import org.wso2.siddhi.core.stream.input.source.InputEventHandlerImpl;
